pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scene_builder"
version = "0.0.1"
authors = [{ name = "SceneBuilder AI", email = "contact@example.com" }]
description = "A system for creating 3D scenes using vision-language models and agentic AI workflows."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "pyyaml",
  "rich",
  "objaverse",
  "pydantic-ai",
  "pydantic-graph",
  "natsort",
  "pillow",
  "scipy",
]

[project.optional-dependencies]
dev = ["pytest"]


[project.urls]
"Homepage" = "https://github.com/example/SceneBuilder"
"Bug Tracker" = "https://github.com/example/SceneBuilder/issues"

[tool.hatch.build.targets.wheel]
packages = ["scene_builder"]
